org: move all utils to backend/pkg/utils
