+++ backend/internal/api/handlers/auth_handler.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/tajjjjr/social-network/backend/internal/models"
 	"github.com/tajjjjr/social-network/backend/internal/service"
-	"github.com/tajjjjr/social-network/backend/utils"
+	"github.com/tajjjjr/social-network/backend/pkg/utils"
 )
 
 // AuthHandler handles HTTP requests for authentication.
