+++ backend/internal/api/handlers/tests/auth_handler_signup_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/tajjjjr/social-network/backend/internal/api/handlers"
 	"github.com/tajjjjr/social-network/backend/internal/models"
-	"github.com/tajjjjr/social-network/backend/utils"
+	"github.com/tajjjjr/social-network/backend/pkg/utils"
 	"golang.org/x/crypto/bcrypt"
 )
