+++ backend/pkg/utils/image_validation.go
