+++ backend/pkg/utils/response.go
@@ -5,10 +5,6 @@ import (
 	"net/http"
 )
 
-type Response struct {
-	Message string `json:"message,omitempty"`
-}
-
 func RespondJSON(w http.ResponseWriter, status int, data interface{}) {
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(status)
