--- backend/utils/types.go
@@ -1,5 +1,0 @@
-package utils
-
-type UserID string
-
-var User_id UserID = "userID"
